+++ internal/deploy/helm/helm.go
@@ -1,0 +1,12 @@
+package helm
+
+import "helm.sh/helm/v3/pkg/chart"
+
+type Deployment struct {
+	Chart  *chart.Chart
+	Values map[string]interface{}
+}
+
+type Deployments struct {
+	Cilium Deployment
+}